+++ packages/icons/CheckBox.svelte
@@ -1,0 +1,28 @@
+<script>
+  export let size = 24;
+</script>
+
+<svg
+  width={size}
+  height={size}
+  viewBox="0 0 18 18"
+  version="1.1"
+  fill="currentColor"
+  xmlns="http://www.w3.org/2000/svg"
+  xmlns:xlink="http://www.w3.org/1999/xlink"
+  xml:space="preserve"
+  xmlns:serif="http://www.serif.com/"
+  style="fill-rule:evenodd;clip-rule:evenodd;stroke-linejoin:round;stroke-miterlimit:2;">
+  <g transform="matrix(1,0,0,1,-3,-3)">
+    <path
+      d="M21,4.474C21,3.66 20.34,3 19.526,3L4.523,3C4.119,3 3.732,3.16
+      3.446,3.446C3.16,3.732 3,4.119 3,4.523L3,19.534C3,19.923 3.154,20.296
+      3.429,20.571C3.704,20.846 4.077,21 4.466,21L19.503,21C20.33,21 21,20.33
+      21,19.503L21,4.474ZM10.336,16.664C10.247,16.753 10.126,16.803
+      10,16.803C9.874,16.803 9.753,16.753 9.664,16.664C8.543,15.543 5,12
+      5,12L6.41,10.59C6.41,10.59 8.656,12.829 9.595,13.766C9.819,13.989
+      10.181,13.989 10.405,13.765C11.977,12.193 17.59,6.58 17.59,6.58L19,8C19,8
+      11.968,15.032 10.336,16.664Z"
+      style="fill:currentColor;" />
+  </g>
+</svg>