+++ packages/icons/base.js
@@ -1,59 +1,41 @@
-import Close from "./Close.svelte";
 import ArrowheadDown from "./ArrowheadDown.svelte";
-import ArrowheadUp from "./ArrowheadUp.svelte";
 import ArrowheadLeft from "./ArrowheadLeft.svelte";
 import ArrowheadRight from "./ArrowheadRight.svelte";
+import ArrowheadUp from "./ArrowheadUp.svelte";
+import BarChart from "./BarChart.svelte";
+import Cancel from "./Cancel.svelte";
+import Check from "./Check.svelte";
+import CheckBox from "./CheckBox.svelte";
+import CheckBoxBlank from "./CheckBoxBlank.svelte";
+import Close from "./Close.svelte";
 import Forward from "./Forward.svelte";
 import Help from "./Help.svelte";
 import Info from "./Info.svelte";
 import Information from "./Information.svelte";
+import LineChart from "./LineChart.svelte";
 import Menu from "./Menu.svelte";
 import Search from "./Search.svelte";
-import Warning from "./Warning.svelte";
-import Check from "./Check.svelte";
-import Cancel from "./Cancel.svelte";
-
-import LineChart from "./LineChart.svelte";
-import BarChart from "./BarChart.svelte";
 import Table from "./Table.svelte";
-
-const icons = {
-  Close,
-  ArrowheadUp,
-  ArrowheadDown,
-  ArrowheadLeft,
-  ArrowheadRight,
-  Forward,
-  Help,
-  Info,
-  Information,
-  Menu,
-  Search,
-  Warning,
-  Check,
-  Cancel,
-  BarChart,
-  Table,
-  LineChart,
-};
-export default icons;
+import Warning from "./Warning.svelte";
 
 export {
-  Close,
-  ArrowheadUp,
   ArrowheadDown,
   ArrowheadLeft,
   ArrowheadRight,
+  ArrowheadUp,
+  BarChart,
+  Cancel,
+  Check,
+  CheckBox,
+  CheckBoxBlank,
+  Close,
   Forward,
   Help,
   Info,
   Information,
+  LineChart,
   Menu,
   Search,
-  Warning,
-  Check,
-  Cancel,
-  BarChart,
   Table,
-  LineChart,
+  Warning,
 };