+++ packages/menu/base.js
@@ -1,0 +1,5 @@
+import FloatingMenu from "./FloatingMenu.svelte";
+import MenuList from "./MenuList.svelte";
+import MenuListItem from "./MenuListItem.svelte";
+
+export { FloatingMenu, MenuList, MenuListItem };