+++ packages/optionmenu/base.js
@@ -1,0 +1,3 @@
+import OptionMenu from "./OptionMenu.svelte";
+
+export { OptionMenu };